+++ src/sqlacodegen/generators.py
@@ -1306,6 +1306,11 @@ class DeclarativeGenerator(TablesGenerator):
             RelationshipType.MANY_TO_ONE,
         ):
             relationship_type = f"'{relationship.target.name}'"
+            if relationship.constraint and any(
+                col.nullable for col in relationship.constraint.columns
+            ):
+                self.add_literal_import("typing", "Optional")
+                relationship_type = f"Optional[{relationship_type}]"
         elif relationship.type == RelationshipType.MANY_TO_MANY:
             self.add_literal_import("typing", "List")
             relationship_type = f"List['{relationship.target.name}']"